refactor: moves available size tests to new file
