+++ test/unit/services/storage/models.spec.ts
@@ -418,23 +418,5 @@ describe('Models', () => {
       ]
       expect(offers).to.be.deep.equal(expectedRes)
     })
-    it('should retrieve min/max value for availableCapacity', async () => {
-      await Offer.bulkCreate([
-        { provider: 'provider1', totalCapacity: 100, peerId: '1' },
-        { provider: 'provider2', totalCapacity: 100, peerId: '2' }
-      ])
-      await Agreement.bulkCreate([
-        { agreementReference: '123', size: 10, offerId: 'provider1' },
-        { agreementReference: '1234', size: 10, offerId: 'provider1' }
-      ])
-
-      const availableCapacityService = new AvailableCapacityService({ Model: Offer })
-      const expectedRes = {
-        min: 80,
-        max: 100,
-      }
-      const minMax = await availableCapacityService.get()
-      expect(minMax).to.be.deep.equal(expectedRes)
-    })
   })
 })
